test_str_get_stringarray_multiple_nans

+++ pandas/tests/strings/test_string_array.py
@@ -101,5 +101,5 @@ def test_string_array_extract(nullable_string_dtype):
 def test_str_get_stringarray_multiple_nans(nullable_string_dtype):
     s = Series(pd.array(["a", "ab", pd.NA, "abc"], dtype=nullable_string_dtype))
     result = s.str.get(2)
-    expected = Series(pd.array([pd.NA, pd.NA, pd.NA, "c"]))
+    expected = Series(pd.array([pd.NA, pd.NA, pd.NA, "c"], dtype=nullable_string_dtype))
     tm.assert_series_equal(result, expected)
